refactor(30-02): extract identity verification queryOptions
- Create identity-verification-keys.ts with queries factory
- Move identityVerificationKeys to query-keys file
- Simplify useIdentityVerificationStatus hook
- Include prior branch work for clean typecheck
