+++ src/components/maintenance/vendors-page.client.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import { useDebouncedCallback } from '#hooks/use-debounced-callback'
 import { Building2, Phone, Mail, DollarSign, Search, Trash2 } from 'lucide-react'
 import { useVendors, useDeleteVendorMutation } from '#hooks/api/use-vendor'
-import type { Vendor, VendorFilters } from '#hooks/api/use-vendor'
+import type { Vendor, VendorFilters } from '#types/domain'
 import { VendorFormDialog } from '#components/maintenance/vendor-form-dialog'
 import { Input } from '#components/ui/input'
 import { Button } from '#components/ui/button'
