+++ src/hooks/api/use-dashboard-hooks.ts
@@ -12,16 +12,13 @@
  */
 
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
-import { createClient } from '#lib/supabase/client'
-import { getCachedUser } from '#lib/supabase/get-cached-user'
-import { handlePostgrestError } from '#lib/postgrest-error-handler'
-import type { FinancialMetrics } from '#types/core'
 import {
-	ownerDashboardKeys,
 	DASHBOARD_BASE_QUERY_OPTIONS,
+	dashboardFinancialQueries,
 	type DashboardStatsData,
 	type DashboardChartsData,
 	type DashboardActivityData,
+	type FinancialTimeRange,
 	type OwnerDashboardData
 } from './use-owner-dashboard'
 import type { PropertyPerformance } from '#types/core'
@@ -128,21 +125,8 @@ export function usePropertyPerformance() {
 // FINANCIAL HOOKS (Revenue/Expense Charts)
 // ============================================================================
 
-export interface FinancialChartDatum {
-	date: string
-	revenue: number
-	expenses: number
-	profit: number
-}
-
-export type FinancialTimeRange = '7d' | '30d' | '6m' | '1y'
-
-const timeRangeToMonths: Record<FinancialTimeRange, number> = {
-	'7d': 1,
-	'30d': 1,
-	'6m': 6,
-	'1y': 12
-}
+// Re-export types for existing consumers
+export type { FinancialChartDatum, FinancialTimeRange } from './use-owner-dashboard'
 
 /**
  * Revenue/expense chart data fetched from the financial analytics RPC.
@@ -150,42 +134,6 @@ const timeRangeToMonths: Record<FinancialTimeRange, number> = {
  * actual expenses instead of placeholders.
  */
 export function useFinancialChartData(timeRange: FinancialTimeRange = '6m') {
-	const months = timeRangeToMonths[timeRange] ?? 6
-	const currentYear = new Date().getFullYear()
-
-	return useQuery({
-		queryKey: [
-			...ownerDashboardKeys.financial.revenueTrends(currentYear),
-			timeRange,
-			months
-		] as const,
-		queryFn: async (): Promise<FinancialChartDatum[]> => {
-			const supabase = createClient()
-			const user = await getCachedUser()
-			if (!user) throw new Error('Not authenticated')
-
-			const { data, error } = await supabase.rpc(
-				'get_revenue_trends_optimized',
-				{ p_user_id: user.id, p_months: 12 }
-			)
-			if (error) handlePostgrestError(error, 'analytics')
-
-			if (!Array.isArray(data) || data.length === 0) return []
-
-			const trimmed = (data as FinancialMetrics[])
-				.sort((a, b) => a.period.localeCompare(b.period))
-				.slice(-months)
-
-			return trimmed.map(item => ({
-				date: item.period,
-				revenue: item.revenue ?? 0,
-				expenses: item.expenses ?? 0,
-				profit: item.netIncome ?? (item.revenue ?? 0) - (item.expenses ?? 0)
-			}))
-		},
-		staleTime: 2 * 60 * 1000,
-		gcTime: 10 * 60 * 1000,
-		structuralSharing: true
-	})
+	return useQuery(dashboardFinancialQueries.chartData(timeRange))
 }
