+++ src/hooks/api/use-owner-dashboard.ts
@@ -10,7 +10,7 @@ import { createClient } from '#lib/supabase/client'
 import { getCachedUser } from '#lib/supabase/get-cached-user'
 import { handlePostgrestError } from '#lib/postgrest-error-handler'
 import type { ActivityItem } from '#types/activity'
-import type { PropertyPerformance } from '#types/core'
+import type { FinancialMetrics, PropertyPerformance } from '#types/core'
 import type { DashboardStats } from '#types/stats'
 import type { MetricTrend, TimeSeriesDataPoint } from '#types/analytics'
 import { revenueTrendsQuery, occupancyTrendsQuery } from './query-keys/analytics-keys'
@@ -42,7 +42,9 @@ export const ownerDashboardKeys = {
 		billingInsights: () =>
 			[...ownerDashboardKeys.financial.all(), 'billing-insights'] as const,
 		revenueTrends: (year: number) =>
-			[...ownerDashboardKeys.financial.all(), 'revenue-trends', year] as const
+			[...ownerDashboardKeys.financial.all(), 'revenue-trends', year] as const,
+		chartData: (year: number, timeRange: string, months: number) =>
+			[...ownerDashboardKeys.financial.all(), 'revenue-trends', year, timeRange, months] as const
 	},
 
 	// Maintenance section
@@ -231,4 +233,61 @@ export const DASHBOARD_BASE_QUERY_OPTIONS = {
 	refetchIntervalInBackground: false,
 	structuralSharing: true
 } as const
+
+// ============================================================================
+// FINANCIAL CHART TYPES & QUERY OPTIONS
+// ============================================================================
+
+export interface FinancialChartDatum {
+	date: string
+	revenue: number
+	expenses: number
+	profit: number
+}
+
+export type FinancialTimeRange = '7d' | '30d' | '6m' | '1y'
+
+const timeRangeToMonths: Record<FinancialTimeRange, number> = {
+	'7d': 1,
+	'30d': 1,
+	'6m': 6,
+	'1y': 12
+}
+
+export const dashboardFinancialQueries = {
+	chartData: (timeRange: FinancialTimeRange = '6m') => {
+		const months = timeRangeToMonths[timeRange] ?? 6
+		const currentYear = new Date().getFullYear()
+		return queryOptions({
+			queryKey: ownerDashboardKeys.financial.chartData(currentYear, timeRange, months),
+			queryFn: async (): Promise<FinancialChartDatum[]> => {
+				const supabase = createClient()
+				const user = await getCachedUser()
+				if (!user) throw new Error('Not authenticated')
+
+				const { data, error } = await supabase.rpc(
+					'get_revenue_trends_optimized',
+					{ p_user_id: user.id, p_months: 12 }
+				)
+				if (error) handlePostgrestError(error, 'analytics')
+
+				if (!Array.isArray(data) || data.length === 0) return []
+
+				const trimmed = (data as FinancialMetrics[])
+					.sort((a, b) => a.period.localeCompare(b.period))
+					.slice(-months)
+
+				return trimmed.map(item => ({
+					date: item.period,
+					revenue: item.revenue ?? 0,
+					expenses: item.expenses ?? 0,
+					profit: item.netIncome ?? (item.revenue ?? 0) - (item.expenses ?? 0)
+				}))
+			},
+			staleTime: 2 * 60 * 1000,
+			gcTime: 10 * 60 * 1000,
+			structuralSharing: true
+		})
+	}
+}
