+++ src/hooks/api/use-owner-notification-settings.ts
@@ -1,17 +1,22 @@
 import { useMutation, useQuery, useQueryClient, mutationOptions } from '@tanstack/react-query'
-import type { NotificationChannelPreferences } from '#types/notifications'
 import type { Database } from '#types/supabase'
 
 import { createClient } from '#lib/supabase/client'
 import { getCachedUser } from '#lib/supabase/get-cached-user'
 import { mutationKeys } from './mutation-keys'
-import { QUERY_CACHE_TIMES } from '#lib/constants/query-config'
 import {
 	handleMutationError,
 	handleMutationSuccess
 } from '#lib/mutation-error-handler'
+import {
+	ownerNotificationSettingsKeys,
+	ownerNotificationSettingsQueries,
+	mapDbRowToPreferences,
+	type OwnerNotificationSettings
+} from './query-keys/owner-notification-settings-keys'
 
-export type OwnerNotificationSettings = NotificationChannelPreferences
+export { ownerNotificationSettingsKeys, type OwnerNotificationSettings } from './query-keys/owner-notification-settings-keys'
+
 export type OwnerNotificationSettingsUpdate = Partial<
 	Omit<OwnerNotificationSettings, 'categories'>
 > & {
@@ -20,36 +25,6 @@ export type OwnerNotificationSettingsUpdate = Partial<
 
 type NotificationSettingsRow =
 	Database['public']['Tables']['notification_settings']['Row']
-
-function mapDbRowToPreferences(
-	row: NotificationSettingsRow
-): OwnerNotificationSettings {
-	return {
-		email: row.email,
-		inApp: row.in_app,
-		push: row.push,
-		sms: row.sms,
-		categories: {
-			maintenance: row.maintenance,
-			leases: row.leases,
-			general: row.general
-		}
-	}
-}
-
-const defaultPreferences: OwnerNotificationSettings = {
-	email: true,
-	inApp: true,
-	push: true,
-	sms: false,
-	categories: {
-		maintenance: true,
-		leases: true,
-		general: true
-	}
-}
-
-const notificationSettingsKey = ['owner', 'notification-settings'] as const
 
 // ============================================================================
 // MUTATION OPTIONS FACTORY
@@ -98,28 +73,7 @@ const ownerNotificationSettingsMutationFactories = {
 // ============================================================================
 
 export function useOwnerNotificationSettings() {
-	return useQuery({
-		queryKey: notificationSettingsKey,
-		queryFn: async (): Promise<OwnerNotificationSettings> => {
-			const supabase = createClient()
-			const user = await getCachedUser()
-
-			if (!user) throw new Error('Not authenticated')
-
-			const { data, error } = await supabase
-				.from('notification_settings')
-				.select('*')
-				.eq('user_id', user.id)
-				.maybeSingle()
-
-			if (error) throw error
-
-			if (data === null) return defaultPreferences
-
-			return mapDbRowToPreferences(data)
-		},
-		...QUERY_CACHE_TIMES.DETAIL
-	})
+	return useQuery(ownerNotificationSettingsQueries.detail())
 }
 
 export function useUpdateOwnerNotificationSettingsMutation() {
@@ -128,15 +82,15 @@ export function useUpdateOwnerNotificationSettingsMutation() {
 	return useMutation({
 		...ownerNotificationSettingsMutationFactories.update(),
 		onMutate: async updates => {
-			await queryClient.cancelQueries({ queryKey: notificationSettingsKey })
+			await queryClient.cancelQueries({ queryKey: ownerNotificationSettingsKeys.all })
 
 			const previous = queryClient.getQueryData<OwnerNotificationSettings>(
-				notificationSettingsKey
+				ownerNotificationSettingsKeys.all
 			)
 
 			if (previous) {
 				queryClient.setQueryData<OwnerNotificationSettings>(
-					notificationSettingsKey,
+					ownerNotificationSettingsKeys.all,
 					(
 						old: OwnerNotificationSettings | undefined
 					): OwnerNotificationSettings | undefined => {
@@ -165,20 +119,20 @@ export function useUpdateOwnerNotificationSettingsMutation() {
 		},
 		onError: (error, _variables, context) => {
 			if (context?.previous) {
-				queryClient.setQueryData(notificationSettingsKey, context.previous)
+				queryClient.setQueryData(ownerNotificationSettingsKeys.all, context.previous)
 			}
 
 			handleMutationError(error, 'Update notification settings')
 		},
 		onSuccess: data => {
-			queryClient.setQueryData(notificationSettingsKey, data)
+			queryClient.setQueryData(ownerNotificationSettingsKeys.all, data)
 			handleMutationSuccess(
 				'Update notification settings',
 				'Your notification preferences have been saved'
 			)
 		},
 		onSettled: () => {
-			queryClient.invalidateQueries({ queryKey: notificationSettingsKey })
+			queryClient.invalidateQueries({ queryKey: ownerNotificationSettingsKeys.all })
 		}
 	})
 }
