+++ src/hooks/api/use-sessions.ts
@@ -2,26 +2,13 @@ import { useMutation, useQuery, useQueryClient, mutationOptions } from '@tanstac
 
 import { createClient } from '#lib/supabase/client'
 import { mutationKeys } from './mutation-keys'
-import { QUERY_CACHE_TIMES } from '#lib/constants/query-config'
 import {
 	handleMutationError,
 	handleMutationSuccess
 } from '#lib/mutation-error-handler'
+import { sessionKeys, sessionQueries, type UserSession } from './query-keys/session-keys'
 
-export interface UserSession {
-	id: string
-	user_id: string
-	created_at: string
-	updated_at: string
-	user_agent: string | null
-	ip: string | null
-	browser: string | null
-	os: string | null
-	device: string | null
-	is_current: boolean
-}
-
-const sessionsKey = ['user', 'sessions'] as const
+export { sessionKeys, type UserSession } from './query-keys/session-keys'
 
 // ============================================================================
 // MUTATION OPTIONS FACTORY
@@ -60,39 +47,7 @@ const sessionMutationFactories = {
 // ============================================================================
 
 export function useUserSessions() {
-	return useQuery({
-		queryKey: sessionsKey,
-		queryFn: async (): Promise<UserSession[]> => {
-			// The Supabase Admin API for listing all user sessions requires the service_role key,
-			// which cannot be used in a browser client for security reasons. The browser client
-			// (anon key) only has access to the current session. We return the current session
-			// as a single-item array.
-			const supabase = createClient()
-			const {
-				data: { session }
-			} = await supabase.auth.getSession()
-
-			if (!session) {
-				return []
-			}
-
-			const currentSession: UserSession = {
-				id: session.access_token,
-				user_id: session.user.id,
-				created_at: new Date(session.user.created_at).toISOString(),
-				updated_at: new Date().toISOString(),
-				user_agent: null,
-				ip: null,
-				browser: null,
-				os: null,
-				device: null,
-				is_current: true
-			}
-
-			return [currentSession]
-		},
-		...QUERY_CACHE_TIMES.DETAIL
-	})
+	return useQuery(sessionQueries.list())
 }
 
 export function useRevokeSessionMutation() {
@@ -101,12 +56,12 @@ export function useRevokeSessionMutation() {
 	return useMutation({
 		...sessionMutationFactories.revoke(),
 		onMutate: async sessionId => {
-			await queryClient.cancelQueries({ queryKey: sessionsKey })
+			await queryClient.cancelQueries({ queryKey: sessionKeys.all })
 
-			const previous = queryClient.getQueryData<UserSession[]>(sessionsKey)
+			const previous = queryClient.getQueryData<UserSession[]>(sessionKeys.all)
 
 			if (previous) {
-				queryClient.setQueryData<UserSession[]>(sessionsKey, old =>
+				queryClient.setQueryData<UserSession[]>(sessionKeys.all, old =>
 					old?.filter(session => session.id !== sessionId)
 				)
 			}
@@ -115,7 +70,7 @@ export function useRevokeSessionMutation() {
 		},
 		onError: (error, _variables, context) => {
 			if (context?.previous) {
-				queryClient.setQueryData(sessionsKey, context.previous)
+				queryClient.setQueryData(sessionKeys.all, context.previous)
 			}
 
 			handleMutationError(error, 'Revoke session')
@@ -127,7 +82,7 @@ export function useRevokeSessionMutation() {
 			)
 		},
 		onSettled: () => {
-			queryClient.invalidateQueries({ queryKey: sessionsKey })
+			queryClient.invalidateQueries({ queryKey: sessionKeys.all })
 		}
 	})
 }
